+++ src/EntryRow.java
@@ -2,4 +2,7 @@ public interface EntryRow {
     public String story();
     public String startedAt();
     public String endedAt();
+    public String description();
+    public Boolean isEmpty();
+    public String toString();
 }