+++ src/Log.java
@@ -1,16 +1,21 @@
 import java.time.Duration;
 import java.time.LocalTime;
-public class Log {
+import java.time.format.DateTimeFormatter;
+public class Log implements Comparable<Log> {
     private String description;
     private String storyCode;
     private LocalTime startTime;
     private LocalTime endTime;
     public Log(
-        String storyCode, LocalTime startTime, LocalTime endTime
+        String storyCode,
+        LocalTime startTime,
+        LocalTime endTime,
+        String description
     ) {
         this.storyCode = storyCode;
         this.startTime = startTime;
         this.endTime = endTime;
+        this.description = description;
     }
     public String storyCode() {
         if (storyCode.length() == 0) {
@@ -24,5 +29,32 @@ public class Log {
             throw new IllegalStateException("Duration is negative.");
         }
         return duration;
+    }
+    public LocalTime startTime() {
+        return startTime;
+    }
+    public LocalTime endTime() {
+        return endTime;
+    }
+    public String description() {
+        return description;
+    }
+    @Override
+    public int compareTo(Log otherLog) {
+        return startTime.compareTo(otherLog.startTime());
+    }
+    public boolean overlaps(Log otherLog) {
+        return
+            startTime.isBefore(otherLog.endTime())
+            && endTime.isAfter(otherLog.startTime())
+        ;
+    }
+    public String toString() {
+        return new StringBuilder(storyCode()).append("-")
+            .append(startTime.format(DateTimeFormatter.ofPattern("HH:mm")))
+            .append("-")
+            .append(endTime.format(DateTimeFormatter.ofPattern("HH:mm")))
+            .toString()
+        ;
     }
 }